Modified input declaration in the genom file
Inputs can be either an IDS member (eg IDS:speedRef) or givn by type (eg double monitor).
The user should not have to worry about storing elements in the ids just for argument passing.
Inputs can also be specified in a more logical way in a comma separated list, or with several
input fields as before.

Also added operator==() for Literal and Service::Input because boost python
requires it to be able to export a std::vector<X>