Many big changes to compile viam with openRTM
* Types in Corba: now all types defined by CORBA have a suffix "_Corba"
at the end. Regular types (ie the ones expected by the codels) are defined
in userStruct.h. Before calling the codel, data is copied to the right structs, etc
* New file: convert.cpp: contains all the functions to convert to/from Corba types
(it is mostly complete, should wok in 90 % of the cases)
* added a new keyword in .gnm: native that indicates that this type is already
defined in another lib and should not be redefined in userStruct.h
(however it is necessary to have the definition as CORBA needs the full
type definition)

*Dynamic ports: a new cpp (opposed to CORBA) is created every time a
codel using the port is called.