Handle forward declaration a bit more precisely
It is important to keep the information where a structure was forward
declared for generating the type mappings. Instead of removing the
forward declaration once a type is fully defined, just rename the key
holding the forward declaration to a special name.

Changed type_final(), dotgen_consolidate() and the Tcl generator accordingly.

While here, factorize a few accessors for idltype_s.