Rework the tcl engine: provide access to the .gen data with C functions
Instead of generating Tcl code to export the .gen data structures, it
is simpler to just use C functions accessing the binary data. For
this, reworked the tcl engine.

The tcl engine is now a binary module, and the Tcl interpreter is
created from C code instead of spawning a binary interpreter. The
genom process is fork()ed before creating the interpreter in order to
provide a bit of separation and also provide more flexibility for
implementing other engines.