Add the tcl interactive template - useful for developing templates
The tcl interactive template does nothing but starting the tcl
interactive loop and interpret user commands. The .gen file is
available as for other templates via the dotgen commands and all the
template commands can be used. This is mostly intended for debugging
and developing the other standard genom templates.

During configure, the eltclsh interactive shell package is looked up
and if found it is automatically loaded by the interactive template
upon startup. It can always be required manually anyway (as well as any
other package).