Add the special /*/ comment keyword to define a copyright notice in .gen files
The notice is copied verbatim in the generated files. This means that
files generated by genom have the same license as the .gen file.

The syntax of the special comment is the same as a 'slash star' C comment:
/*/ Text ...
 ... can span multiple lines */
Only the first block appearing in the .gen file is considered.