Store integer IDL constants in a 64 bits wide integer.
This is to handle the upcoming long long types.

The value of IDL constants declared in the .gen are only used internally; they
are passed to the TCL engine via ASCII strings. So their actual storage size is
the same for all intger constants, and increasing this to 64 bits is harmless.