Implement the first draft of the C++ mappings
For bounded version of string and sequences, the mapping relies on a runtime
header that contains a generic implementation of those types. Both are structs
with a public array of fixed size (a bit like tr1/array, unfortunately not in
C++ yet). For now, those struts provide no method, until we decide what is
really needed.

Unbounded strings and sequences use the STL string and vector types directly
(since having a pointer here is not so much a penalty: some kind of
serialization will have to be done anyway).

Unions are mapped to a struct containing the discriminant plus a native union
(much like the C mapping). Union members whose C++ mapping have a non-trivial
constructor are not allowed. This is a consequence of the C++ std. The
forbidden types are sequences and strings (bounded or not), and any structure
or union that contain those. This introduces a disymetry with the C mapping,
but since such weird unions are probably useless, it's better to keep the union
mapping simple and not go toward such an awful mapping as in CORBA (IMHO).

The other (basic) types have the same mapping as in C. Integer types use
stdint.h (not cstdint as this is not part of C++ yet).