Implement components 'attributes'.
Component attributes are values from the IDS that are considered as global
parameters controlling the component behaviour. This can typically be a driver
path or a parameter file name, aso.

genom does not further refine the meaning of attributes, but this can for
instance be mapped in a template into automatically generated getters and
setters, or some request to save and later reload those attributes.
In an Orocos template, this would typically be mapped into orocos
'properties'.

In tcl engine, add [$component attribute] to retrive attributes as a list of
parameters pointing to members of the IDS.