Fix skeleton template with c++ component and add -l switch for codels language
The -l switch selects the codel language. This is not to be confused with the
codel interface language defined in the .gen. If the language defined with -l
is different from the language defined in the component, only one combination
is supported at the moment: c interface with c++ codels. This corresponds to
what genom2 implements.