Add type_renew() that makes a type appear as if it had just been defined.
Hash tables storing types are ordered, so that type definitions appear in the
right order when iterating over them. type_renew() update the type hash tables
so that a previously defined type appear as if it had just been defined, so
after any other existing type.

This is useful for an upcoming commit allowing to extend the ids definition.