Ids definition is now done a the global level (not inside component object).
Instead of simply assigning a type to the ids inside a component, the ids is
now a top-level genom object and is declared much like an IDL structure, like
this:

ids {
    long a;
    <type> <member>;
    ...
}

Of course, a component must have previously declared. Several ids statements
are allowed, and the resulting ids is the concatenation of all the defined
structures.

The motivation for this change is to allow to build the ids in several
chunks. This is especially useful for templates that automatically add new
standard services to a .gen description and let them decalre new input/outputs
in the ids.