Alias .gen object command in the slave template interpreter directly in C code.
Each .gen object is mapped to a tcl procedure in the main tcl engine
interpreter. It used to be aliased in the slave interpreter that parse template
files from TCL code, but it is easier to do so from the C code that creates the
TCL procedure, so that the [dotgen parse] procedure automatically exports newly
created object in the slave as well.