Add a hash in tasks and services that indexes the codels from the 'events' pov
New internal functions task_fsm() and service_fsm() retrive the task and
service fsm hash. The fsm hash is indexed by events name and contain the codel
associated to the event.

Task and service objects in TCL engine now accept the 'fsm' method that return
the codel object associated to the event passed as a parameter.

While here, check that services and tasks associate only one codel to an event
and report an error during parsing otherwise.