Rewrite tcl engine "language declarator", add "parameter" and "argument" procs
To clean up the template procedures related to parameters, arguments and
declarations, the procedure "declarator*" and "declarator&" have been deleted.

We have now only one way to declare a type or a variable of a type with
"language declarator". Declaring a pointer or a reference can be done by
passing a variable of the form *var or &var, but this is not handled specially
by "language declarator" and it is of course language-specific.

Declaring a parameter of a function can be done with "language parameter"
(which ouputs something like "int a" or "int *a" in C). The procedure is passed
an argument that can be either "value" or "reference" to declare the parameter
to be passed by value or by reference, whatever this means for a particular
language.

Passing an argument to a function can be done with "language argument" (which
outputs something like "a" or "&a" in C). It is also passed a "value" or
"reference" argument.

"language parameter" and "language argument" handle correctly the type of the
argument (e.g. strings or sequences), to pass them by value or reference. (in C
or C++ atm).