Change the skeleton template so that it generates an installable lib for codels
While trying to make genom not impose any build system for the templates, it
was necessary to rework the skeleton template. The idea is to only impose the
presence of a pkg-config file (.pc) that describes how to link with the codels
and that will be used by the templates.

With this commit, the skeleton template now generates a complete example source
tree with empty codels that can be filled later. The generated files compile a
library lib$compenent_codels that should be installed before trying to compile
any template. Thus, to compile a useful genom3 module, one will have to first
compile/install codels and then generate/configure/compile one or more
templates. This is a bit more complex than with genom2, but this is basically
the price to pay to let templates use the build system of their choice. A
wrapper script/makefile may be developped later to simplify this process.