Add a new "port array" type
The 'port array' is meant to be used by components that create a variable
number of posters, whose name or quantity cannot be determined at compile time.

Port arrays are declared by appending an empty `[]' right after the port
declaration, like in "outport data<type> name[];". Any reference to this port
in codel parameters must specify an IDS member that is used to compute the name
of an element of the poster array. For example: outport name[idsmember]. The
IDS member can be of any type, but must be a scalar or a string.

This is not compatible with the 'scheduling: real-time' property,
however no check is implemented yet. Time will tell if this should be added or
not, perhaps by shifting the 'real-time' property to more specific elements of
a component (like a per-task attribute).