Add the 'merge-if-change' engine mode and 'engine merge-tool' procedure
In addition to the 'move-if-change' and 'overwrite' file generation mode, the
new 'merge-if-change' mode invokes a merge procedure if a generated file
already exists but would be modified (overwritten) by the engine.

For that, the 'engine merge-tool' procedure may be used to define a merging
tool. It can be either the builtin 'interactive' or 'auto' keywords, or the
path to an executable.

'interactive' or 'auto' invoke a builtin procedure that performs automatic
merging of the two versions of the file. The only difference between
'interactive' and 'auto' is that in case of a conflict, 'interactive' will ask
the user (on stdin) for each patch hunk resulting from the diff on the two
versions if it should be applied, while 'auto' will place a conflict marker
with both versions in the generated file.

If the merge tool is set to an executable, this program is invoked with the two
versions of the file. The tool can be anything that accepts two files on the
command line, such as 'meld'. The tool is expected to save the new version of
the file: no further action is taken by the engine itself.