Update engine's verbosity control
Engine verbosity used to be controlled by redirecting stdout to one pipe or
another in the parent process. This raises an issue in the 'interactive'
template and recent libedit versions, because recent libedit refuse to turn
interactive editing mode on on something that is not a tty.

Update the strategy to send engine's output to it true 'stdout' in verbose
mode, and a pipe to the parent in silent mode. All in all this simplifies a bit
this messy part ...