Massive rewrite of the "language" engine procedure
All the "language" engine procedures have been simplified to reduce
typing. All procedures working on a genom object type (codel, idl type,
parameter...) have been converted to an OO version, and a new "lang" procedure
can be used once for all to configure the programming language of the generated
strings. The $type foreach procedure has also been improved and specialized to
generate array and sequence loops automatically in the selected language (based
on the first few templates experience, this is what is the most needed). Check
the documentation for up-to-date information.

No functional change (intended ...).