Round period in microseconds to the nearest integer for default clock_rate
The default clock_rate (when none is provided in a .gen) is computed as the GCD
of all periods/delays in the tasks, with a microsecond resolution. The
conversion from double to integer may cause problems (real-life example: 30ms
rounded to 29999us), so make sure to round to the nearest integer by
adding 0.5 to the microseconds in double format.