Build LIBOBJS compat functions as a standalone convenience library
The compatibitly functions are now used by genom and genom-pcpp from two
different directories. Automake miserabily fails to handle this: the distclean
target removes compat/.deps/*.Po files too early and breaks other Makefiles.

A workaround is to build LIBOBJS as a standalone convenience library. As
mentionned in the automake documentation, one has to avoid building an empty
library if LIBOBJS is empty. For this, the xwarnx() function was elected as the
one to move from src/ to compat/.