Switch to using genom-pcpp as the default C preprocessor for .gen files.
Remove checks for cpp and -xc required option for gcc -E in configure script.
Compile default libexec/genom-pcpp default value in genom. This is still
overridable with CPP env. variable  at runtime.

One side effect is the removal on an extraneous whitespace in the regression
test for cpp. This extra space was generated by gcc -E, but not by pcpp.