Implement "#pragma require" feature in genom-pcpp
genom-pcpp now recognizes the #pragma require <pkg> directive. The effect of
this directive is to lookup <pkg> using pkg-config (obeying PKG_CONFIG and
PKG_CONFIG_PATH env. vars) and dynamically add any -I and -D directive found in
the Cflags of pkg, before continuing parsing the file.

Multiple <pkg> may be specified in a single #pragma require, and <pkg> should
be double quoted if it contains spaces (as in "x > 1").

This is intended to solve the chicken-egg problem of the genom "require:" field
of components, where the "require" in question must be known before parsing the
file (typically because of #include'd files).