Silent a gcc-4.5 warning about a 'case' label not in enumerated type
The enum in question is a set of flags that are or'ed together, so obviously
the or'ed combination should not be defined in the enum (otherwise there would
be no point in or'ing anything).
Anyway ... casted the enum to unsigned to silent -Wswitch with gcc-4.5

See http://gcc.gnu.org/bugzilla/show_bug.cgi?id=46457 for a similar complaint.