Add a 'inport handle' kind of inport, delete 'inport event' and 'inport arrays'
In order to rationalize things before 3.0, delete unused 'inport event' kind of
inports. This won't be in 3.0.

Also remove 'inport arrays' for dynamic port creation, and replace by a 'inport
handle' that will be used in a genom port API implemented by the templates, to
grant unlimited access to in/outports to the codels. This kind of access won't
of course be formal... Two kinds of handle are provided: scalar and sequence
(former ports arrays).

While here, delete 'data' keyword in '{in,out}port data<type>', so that regular
ports are now defined by '{in,out}port<type>' and handle ports by
'{in,out}port handle<type>'.