Update parameters "type" member to reflect the actual type of the parameter
The "type" parameter member could refer to the struct member of the IDS or a
nested struct, but this is rather painful when writing templates and was not
really useful. So update the code to always set the "type" member to the actual
IDL type object of the parameter. The member name (if any) is still accessible
through the "member" parameter member.