Update again IDL/C mapping for enums
Defining enumerators as "static const uint32_t" does not work when the
enumerator is to be used as a static array initializer.

So, to work around all issues, define the enumerated type as uint32_t (so that
sizeof is guaranteed to be 4) and define the enumerators in a real (anonymous)
enum.