Fix an issue with passing multidimensional arrays by value (const)
For the specific case of IDL arrays of arrays or arrays of string, the C
language requires a cast to properly pass a non-const parameter as a const
argument of a function. Since C does not allow cast-to-arrays either, one must
fiddle with the cast argument and integrate the decay-to-pointer of arrays in
the cast.

See http://c-faq.com/ansi/constmismatch.html for a comprehensive description of
the problem.