Add the generic client interface header in the runtime API
This header is not strictly a 'runtime' file, but it is useful to compile
any runtime client. It is also not really required by genom, since genom does
not impose anything on clients, but putting it here makes it clear that the
client API is template independent (and making a standalone package with only
this file is a bit overkill).