Make skeleton template generate a configure script handling --with-templates
The --with-templates option takes a comma separated list of templates that
should be generated/configured and compiled along with the codels. At the
moment it only supports autoconf style templates. Configure options passed to
the codels are recursively passed down to templates (as with AC_CONFIG_SUDIRS).

To deal with this, an extra libcodels-uninstalled.pc file is provided (never
installed) so that "server" templates can link with uninstalled codels.

Note: this feature is by nature not robust as it cannot easily handle templates
specificities. It is provided only as a helper for most common situations. For
fine tuning template generation and compilation, the manual approach should be
preferred.