Let eng_listtmpl() return an array of strings instead of printing the result
This makes eng_listtmpl() more generic and it will be useable in the tcl
backend. Transfer the actual printing of the list of templates into main().
While here, remove the printing of the template backend name, this is of little
interest (especially since there is only one backend for now ;).

Both strings and the array must be freed by the caller.