Better report error messages from merge tools during a merge in tcl backend
The 'exec' tcl command considers by default that an application writing to
stderr is an error condition. Since we don't want this behaviour and we're also
interested in live output during an interactive merge, redirect application's
channels to the backend's ones (fixes merges with 'meld' that spits a few gtk
warnings on stderr).

While here, also catch any (real) error message and report it so that
the user has an idea of what happened instead of a mere "merge failed".