Add an 'async' kind of codels
Async codels are declared with "async codel <event>". The former "codel
<event>" declaration declares a normal, synchronous codel.

Async codel are allowed to block (typically waiting for an internal event on a
sensor, but there is no restriction on the kind of things it may do). When an
async codel runs, other execution contexts (ie tasks) will process their
activities in parallel. The task of the async codel will still be blocked, so
that no two codels of the same task never run in parallel.

Some care must be taken when dealing with output parameters in the IDS: since
other codel may run while the async codel is running, async codels should
probably not have much output and just return a new activity regular state (not
async) that can take care of writing an output to the IDS.