Write the documentation with texinfo instead of LaTeX
LaTeX is great, but it's difficult to make it output a proper html file: tth is
not so bad, but the free version cannot split the html file, which is a bit
painful if the documenation grows significantly (will it?).

texinfo is not great, but it is simple and can output ps, dvi, pdf, html, info,
text and xml with not so much trouble (even docbook, although I could not make
this work...). And it is not obvious that modern alternatives are actually
better than texinfo.

With texinfo, we also get an info doc for free, which is really useful for a
quick lookup.

All the embedded documentation (esp. in dotgen.y and tcl engine) have been
converted, as well as the scripts extracting texinfo bits from the source.