Add param_equal() and param_list_equal() that compare two lists of parameters
param_equal() compares two parameters and returns true if the two refer both to
the IDS or the same port and type_equal() returns true for the underlying
types. (The parameter name is thus irrelevant.)

param_list_equal() returns true if the two lists of parameters are the same
according to param_equal(). (order matters).