Forbid the declaration of codels with conflicting prototypes
Two services can only share the same codel if they define the same signature.

This is so because the underlying C symbol has this restriction. It is not
clear yet how C++ codels will be defined for pure C++ components, nor if this
will ever be implemented, nor if the above restriction will hold, but for now
let's do it like this.