Provide a generic genom.gen file, that templates can use
The generic genom.gen defines a "state" outport (former control poster of
GenoM2) as well as two requests for aborting a service and connecting an inport
to an outport.

The file is stored in [dotgen template builtindir]/common/genom.gen, and can be
passed to [dotgen parse file]. It can be used by templates to provide this
functionality in the components they produce, but this is not mandatory (for
instance the 'skeleton' template does not want to know about these, since user
is not supposed to see this).