Add a utility procedure for computing a MD5 digest of genom objects
This is a nice idea borrowed from ROS that does a kind of automatic versioning
of a component interface. An MD5 hash can be computed on the component events,
ports and services parameters so that if anything in the interface changes (in
an incompatible manner), the hash will be different. The hash is to be
embedded both in server and clients, so that a client can compare it with the
server's one and refuse to talk to strangers.

The MD5 routines are pure tcl borrowed from tcllib and fine tuned to our use
case - because tcllib is not that widespread (not always packaged, up-to-date
or whatever) and the tcllib license is permissive (BSD with no
advertising). While some libc have MD5* routines, this is a bit system
dependant (linux has them in libbsd, darwin in openssl only, *Bsd in libc) and
trying to sort this out is not worth the trouble.

The tcl backend makes the digest routines available as [$object
digest]. Currently, object must be a component or an IDL type object. The
procedure returns a 33 bytes string (32 + nul).