Delete tcl backend [$type foreach] procedure
$type foreach was used to flatten a recursion on an IDL type definition and
execute a user script for each member of a constructed type. The procedure was
complex and hard to maintain and usage was all the more obscure.

The ros and pocolibs templates use cases have shown that it is more
straightforward and elegant to write recursive tcl code also generating
recursive C (or C++) code (possibly "inlined" or "templated" if performance is
really an issue).

So say goodbye to the [$type foreach] attempt. It's still in git anyway.