Delete TCL backend [language iter] procedure
[language iter] was used in combination with [$type foreach] to generate
the source code of iterators on arrays/sequences in the target language.

This was even more obsfuscated than the now obsolete [$type foreach], and usage
has shown that trying to address the issue this way was not the correct
solution. Source code generation really belongs to the template, the backend
must not be aware of this. Goodbye!