Change service parameters semantics: parameters are local to activities
The dotgen grammar has been updated in an incompatible way: each service
parameter now defines a variable local to each instance of the running
service. This is to prevent poorly defined semantics when two concurrent
services wanted to access the same IDS member. As a consequence, regular
services cannot update or read the IDS alone (without the help of a user
codel).

Codels must now explicitly declare their parameters as coming from the service,
IDS or port, by prefixing each parameter with 'service', 'ids' or 'port'. When
there is no ambiguity, this declaration is optional.

For symmetry, 'inport' and 'outport' keywords have been replaced by 'port in'
and 'port out' (2 words), and the syntax of port declarations updated.