Update 'attribute' syntax and semantics
Since services cannot update the IDS alone anymore, the semantics of
'attributes' has been refined so that it is possible to continue writing IDS
accessors in a simple manner. Attributes are declared with the same syntax as
services, except that they take their parameters from the IDS. They can only
have 'validate' codels (i.e no task, no state machine).

Without a validate codel, an attribute will simply store its input in the IDS
and read its output from the IDS. With a validate codel, it will be able to
check input and throw some exception before doing so. In all other aspects
(interruption, before/after, throw...), attributes are similar to services.

Of course, it is the responsability of the templates to implement this
behaviour.