Rename a few badly named symbols in generic C client interface
Rename request_info structure to service_info, since this structure holds the
description of a service, not a request.

Change all references to an "activity" to "rqstid", since activity now refer to
a particular service and "rqstid" is probably more self explanatory anyway.

Rename the opaque struct genom_client to genom_client_s because 'typedef struct
genom_client *genom_client' is not valid in C++.