Implement 'remote' objects declaration and IDL remote type kind
'remote' objects refer to distant services that a codel can use.

Internally, remotes contain a 'remote' IDL type kind. This type is meant to be
used by 'remote' genom objects, so that the mappings of codel parameters into
the programming language is handled consistenly.

The C mapping of such an object is a structure containing a pointer to a 'call'
function, with a prototype corresponding to the 'remote' arguments. The C++
mapping is not defined yet (as many other c++ aspects are still to be
implemented).