Implement basic 'extends', 'provides' and 'uses' properties
'extends' let an interface inherit from another one. For now, this simply
merges the extended interface into the extending one.

'provides' let a component implement an interface. For now, this simply import
all definitions of the interface into the component.

'uses' let a component import all the services defined in the interface as
remote services.

This implementation is sufficient to share service declarations between
'remote' objects and local implementations, so that parameters names are
guaranteed to be the same on both sides (the ROS template requires this).

However, it raises a number of issues that are not solved yet. In particular:
 - port and multiple port definitions?
 - namespace?
Until this is properly considered and solved, interface definitions are not
constrained too much and they can basically declare everything that a component
can define.