Add a list of "common" exceptions to be thrown by templates and std services
Letting templates define their own exceptions raise an issue with the generated
enum implementing the exception list. Due to the way the parser works, the enum
may become desynchronized between the codels and the server implementation.

The easiest workaround is to centralize those common exceptions in genom
itself, so that everybody agrees on these. This has also the advantage of
making error handling in the supervision a bit more idependant of the
underlying templates.