Rewrite the genom_resize_sequence macro to avoid breaking strict aliasing
ISO C forbids accessing an object through something else than a pointer on its
real type, so reallocating an IDL sequence cannot be done by passing a void **
pointer on the buffer and dereferencing the void **. Rewrite the helper
function so that it returns the new buffer address (like realloc()). The helper
macro on top of the helper function masks the details.