Implement "native" type declaration, for use in the IDS
The "native" IDL keyword introduces the definition of a native type. A native
type is opaque to genom and templates and can only be manipulated by user
code. Consequently, it can only appear in IDS and codels, not in services
parameters or ports. Native types may typically be used to store private data
in the IDS, that cannot be represented in IDL (e.g. types defined in external
libraries, opaque types, etc.).

The "native <identifier>;" declaration defines a new native type, named by
<identifier>. The C mapping is a forward declaration of a structure, with the
same scoping and naming rules as for other types. The user is expected to
provide a definition for the structure, although this is not strictly required
as genom and templates will only manipulate a pointer on this structure. The
definition will be required when the codel will want to manipulate the type.
Memory management associated with the pointed data must also be handled by the
user.