Update C mapping of bounded sequences because of a C++ hitch
The C mapping of a bounded sequence contains a 'const uint32_t _maximum'
member, which represents the (fixed) max length of the sequence. This is valid
C, but in order to be valid C++, it requires a constructor with an explicit
initialization of the const member in the initialization list.

Even though the C mappings should not care about C++, there is no other
possible workaround than putting this inside an #idef __cpluscplus during the
C mappings generation.

And even when the C++ mappings will be fully functional, including the C
mappings of IDL types in C++ code is useful at least for:
 - mixing C++ server template with C codels.
 - compiling C codels with a C++ compiler.