Implement proper "exceptions" in lieu of integer reports
Exceptions replace the integer reports of services. This allows services to
report additional information in case of abnormal termination.

In a .gen file, exceptions are defined like a 'struct' but with the
'exception' keyword instead. Empty (void) exceptions can also be defined. This
is in line with the OMG IDL definition of exceptions. The 'throw' keyword must
now be used with such an exception type.

Some incompatible changes in the codels API had to be introduced. The return
type of codels is now "genom_event". All codels transitions and exceptions are
of that type. A codel transition is tiggered by returning a genom_event of the
form <component>_<state> (as before). A codel exception is triggered (in C) by
returning the result of the invocation of a function throwing the exception
(the C++ mapping is expected to use the C++ 'throw' functionality). Stateless
codels (including 'validate') must now return either genom_ok, or an exception.