Improve error reporting in tcl code generation engine
Tcl-8.5 reports only a human-readable backtrace upon error. Within genom, this
backtrace refers to a generated tcl program and not to the actual template
source file, making errors a bit difficult to interpret.

With this commit, a proper backtrace containing line numbers and source code
excerpts is generated. This is implemented by adding an execution trace that
builds the invocation context during stack unwinding. This is quite costly (it
more or less doubles the processing time), but this is barely noticeable in
practice. (Tcl-8.6 has a builtin way of generating programmatically exploitable
backtraces, but until this version is more widespread it's better to stick to
tcl-8.5.)

In order to simplify error line number reporting, the generated program
corresponding to a template source file takes care not to add or remove any
newlines. This way, error lines reported by tcl directly match those of the
source file and can be reported as-is. Also, the program is generated in a
temporary file instead of in-memory string, because error line numbers are more
easily displayed this way (error lines in tcl eval'ed strings are relative,
while they are absolute when processing a file).

Finally, the temporary generated program now uses special commands
corresponding to the template markups so that they can be identified in the
backtrace and converted back to the original markup when reporting errors.