Add an optional 'filter' when computing [$component digest]
The optional filter can be used to filter out some elements from the digest
computation. This is used for instance by the pocolibs template to exclude the
private "metadata" port that is not visible to clients.

The filter must be a tcl anonymous function (see tcl [apply] command) that
accepts one argument that is a genom object. It must return a boolean to
indicate whether the object should be included (true) or excluded (false) from
the digest.