Add [$component types ?visibility? ?filter?] and [$type types]
[$component types ?visibility? ?filter?] returns a list of types defined by a
component. "visibility" must be "public" or "private". The returned list of
types includes either only the publicly visible types (i.e. part or the
services/port interface) or all the types (i.e. the interfaces plus all the
types used in the ids, codels, remotes, etc.).

The optional filter can be used to filter out some elements from the type
list.The filter must be a tcl anonymous function (see tcl [apply] command) that
accepts one argument that is a genom object. It must return a boolean to
indicate whether the type should be included (true) or excluded (false).

[$type types] returns a list of all types defined recursively by the given
type.

These new commands factor out a functionality useful for all templates that
used to be implemented separately by the templates.