Update IDL to C++ mappings
This fixes a number of left out and broken things in the first draft
implementation, and implement missing mappings for the newest constructions.
Note that no template is currently able to use this, and especially no server
template so that codels still have to use the C interface at the moment.

Preliminary template classes for bounded strings and sequences are dropped and
replaced by the use of array of chars (bounded strings) and tr1::array (bounded
sequences; std::array is still not officially supported by most compilers, so
it could not be used).

An implementation of genom::event is provided and is mostly the same as for
the C mappings (a global 'weak' symbol to an array of chars).

Exceptions map onto a real type that can be thrown, inherit of genom::exception
and std::runtime_error and implement a what() method returning the exception
name.

Ports and remotes (for codels) provide a pure virtual interface (that must be
implemented by a template) for accessing the data.

Other types have straightforward mappings and deserve no special comment.