Fix error reporting in templates (once more)
This commit is quite messy, mostly because tcl cannot report a proper backtrace
of error (it reports a formated backtrace, but nothing that can be
programmatically used).

This commit re-implements a "source" command (using Tcl_EvalFile) that does not
add its own error stack/context, so that one can report the error line inside
the sourced file itself (of course, this information is reported by the
vanilla 'source' command in the formatted errorInfo, but it is lost for
instance in the -errorline returned by a catch {source ...}). This is useful
for reporting errors that happen in a "template require" genom procedure.

This commit also fixes a few gory details in errors reported while parsing a
template source file. Sigh :)

BTW: tcl-8.6 is supposed to improve this with the -errorstack information
available in 'catch' contexts, but this was not tested.