Redo the C++ mapping for bounded sequences
std::tr1::array cannot be used for the mapping of bounded sequences, as the
size() of a std::tr1::array is a constant and thus cannot be modified.

So revert to defining a custom template type, namely genom::bounded_sequence<>,
which closely ressembles the std::tr1::array struct ... and is actually mostly
pasted from the llvm array implementation.