Add a "result" function in the generic client/c interface
The "result" function is meant for retrieving the output, report and exception
details of an active request. "result" blocks until the request is terminated.

This updates the semantics of the "send" function a bit (but since it was not
documented yet, this is not an issue ;). "send" is now always asynchronous and
never blocks. If "send" is passed non-NULL callbacks, it behaves as before. If
it is passed NULL callbacks, this simply means that no callback is invoked on
the "sent" and "done" events. The "done" function can be used to query the
current request status in a non-blocking way, and the "result" function can
retrieve the output data.