Fix the generic C++ genom::exception definition.
It is not possible to make the genom::exception extend the STL runtime
exception, because the genom exception has no member and the STL runtime
exception must be initialized with a string. And it was probably not clear to
mix STL exceptions and genom exceptions anyway.

So define a bare struct instead, just for the sake of being able to catch any
genom exception in one statement.

Thanks to adegroot@ for the looooong bug report :)