Add a .pc file for exporting the C client protocol number
The C client protocol number is already checked at runtime, but it is
potentially required at build time (to prevent weird compilation failures and
report errors early/clearly).

The idea is to have the various client/c generate a .pc file for their
instanciated client, requiring genom3-c-client with the exact protocol number
that they are going to generate.