Update directories in common C clients .pc file
Install C client libraries and headers in a middleware-dependent location. This
is needed to install clients for different middlewares in the same prefix.

The name of the libraries and headers are still kept unchanged, though, so that
only a PKG_CONFIG_PATH update is required to link with different clients for
different middlewares.