Fix some objects iterators in the generator so that they stop randomizing lists
Fix the 'throws' method of components/remotes/services/tasks so that the
returned list is kept in a deterministic order. This is important for several
reasons, the most important one being that the client and server md5 digests
rely on it.

The problem was that in order to remove duplicates in the list, the tcl "lsort
-unique" procedure was used. While this actually removes duplicates, this
paradoxically introduces a randomization of the list, because the string
representation of genom objects in Tcl is based on their address in memory.
So just use a trivial duplicates removal that preserves order instead.

Now, why this has never been an issue before is another question, but thanks to
rzhao@ for unveiling this!